Use `NSUInteger` instead of `int` for button states array ivar

+++ TimesSquare/TSQCalendarRowCell.m
@@ -13,7 +13,7 @@
 
 @interface TSQCalendarRowCell ()
 {
-    int buttonStates[7];
+    NSUInteger buttonStates[7];
 }
 
 @property (nonatomic, assign) NSInteger indexOfTodayButton;
